split "util_functions.py"; moved tests and experiments; excluded them from Docker build
