+++ .gitignore
@@ -163,3 +163,4 @@ cython_debug/
 /data
 /tmp
 .idea
+/AGENTS.md
