+++ pyproject.toml
@@ -67,6 +67,7 @@ rag = [
 ]
 test = [
     "black>=25.9.0",
+    "igraph>=1.0.0",
 ]
 
 [tool.uv]
@@ -75,6 +76,10 @@ default-groups = [
     "vectorstore",
     "rag",
 ]
+
+[tool.pytest.ini_options]
+pythonpath = ["."]
+testpaths = ["test"]
 
 [tool.hatch.build.targets.sdist]
 include = [
