+++ src/common/iterables.py
@@ -1,0 +1,9 @@
+"""Iterable helpers."""
+
+import itertools
+
+
+def pairwise(iterable):
+    a, b = itertools.tee(iterable)
+    next(b, None)
+    return zip(a, b)
