+++ src/core/clustering_config.py
@@ -1,0 +1,9 @@
+"""Configuration enums for clustering algorithms."""
+
+import enum
+
+
+class ClusterNumberDetection(enum.Enum):
+    DISTORTION = 1
+    SILHOUETTE = 2
+    CALINSKI_HARABASZ = 3
