+++ src/core/reduction.py
@@ -1,0 +1,9 @@
+"""Dimensionality-reduction helper types."""
+
+
+class NoneDownScaleObj:
+    def __init__(self, **kwargs):
+        pass
+
+    def __str__(self):
+        return "None"
