+++ src/pipeline/steps/embedding_util.py
@@ -7,7 +7,7 @@ from src.pipeline.base import BaseUtil
 from src.pipeline.status import StepsName
 from src.storage.marqo_external_utils import MarqoEmbeddingStore
 from src.core.embedding_functions import SentenceEmbeddingsFactory
-from src.common.util_functions import load_pickle
+from src.common.io import load_pickle
 from src.pipeline.load_utils import FactoryLoader
