--- src/tests/test_data_functions.py
@@ -1,8 +1,0 @@
-from unittest import TestCase
-
-from src.core.data_functions import DataProcessingFactory
-
-
-class TestDataProcessing(TestCase):
-    def test_data_processing_class_is_exposed_by_factory(self):
-        self.assertTrue(hasattr(DataProcessingFactory, "DataProcessing"))
